refactor: move product custom hooks
